+++ docs/changes.rst
@@ -13,6 +13,8 @@ v2.4.3 (UNRELEASED)
 **Crawlers**
 
 - New: ``firekanta``
+- New: ``truthfacts-gocomics``
+- Update: ``lunchtu`` after site change
 
 v2.4.2 (2016-01-19)
 ===================