+++ docs/changes.rst
@@ -12,6 +12,7 @@ v2.4.3 (UNRELEASED)
 
 **Crawlers**
 
+- New: ``businesscat``
 - New: ``firekanta``
 - New: ``mollybeans``
 - New: ``truthfactsgo``