+++ docs/changes.rst
@@ -18,7 +18,8 @@ v2.4.3 (UNRELEASED)
 - New: ``mollybeans``
 - New: ``truthfactsgo``
 - New: ``wyyrd``
-- Update: ``lunchtu`` after site change
+- Update: ``lunchtu`` after site change.
+- Update: ``thegamercat`` after feed change.
 
 v2.4.2 (2016-01-19)
 ===================