+++ docs/changes.rst
@@ -5,7 +5,7 @@ Changes
 This change log is used to track all major changes to *comics* after the first
 versioned release.
 
-v2.4.3 (UNRELEASED)
+v2.4.3 (2017-04-01)
 ===================
 
 - Update all crawlers based on ``HeltNormaltCrawlerBase`` after site change.